create /react subpackage
