+++ examples/with-openai/app/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useChat } from 'ai-connector'
+import { useChat } from 'ai-connector/react'
 
 export default function Chat() {
   const { messages, input, handleInputChange, handleSubmit } = useChat()
