+++ packages/core/package.json
@@ -7,7 +7,8 @@
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
   "files": [
-    "dist/**"
+    "dist/**/*",
+    "dist/react/**/*"
   ],
   "scripts": {
     "build": "tsup",
@@ -17,6 +18,21 @@
     "type-check": "tsc --noEmit",
     "prettier-check": "prettier --check \"src/**/*.ts*\"",
     "test": "jest --env @edge-runtime/jest-environment .test.ts && jest --env node .test.ts"
+  },
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "module": "./dist/index.mjs",
+      "require": "./dist/index.js"
+    },
+    "./react": {
+      "types": "./react/dist/index.d.ts",
+      "import": "./react/dist/index.mjs",
+      "module": "./react/dist/index.mjs",
+      "require": "./react/dist/index.js"
+    }
   },
   "jest": {
     "preset": "ts-jest",
@@ -44,6 +60,11 @@
   },
   "peerDependencies": {
     "react": "^18.0.0"
+  },
+  "peerDependenciesMeta": {
+    "react": {
+      "optional": true
+    }
   },
   "engines": {
     "node": ">=14.6"
