+++ packages/core/react/package.json
@@ -1,0 +1,10 @@
+{
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": "./dist/index.mjs",
+  "private": true,
+  "peerDependencies": {
+    "react": "*"
+  }
+}
