+++ packages/core/streams/anthropic-stream.ts
