+++ packages/core/streams/index.ts
@@ -3,6 +3,5 @@ export * from './openai-stream'
 export * from './streaming-text-response'
 export * from './huggingface-stream'
 export * from './anthropic-stream'
-export * from './use-chat'
-export * from './use-completion'
 export * from './langchain-stream'
+export * from '../shared/types'
