+++ packages/core/streams/openai-stream.ts
