+++ packages/core/streams/streaming-text-response.ts
